+++ MapReduce/src/hdfs/KPFSMasterInterface.java
@@ -15,6 +15,6 @@ public interface KPFSMasterInterface extends Remote {
 	public ArrayList<String> splitFile(String filePath, int chunkSizeB,
 			String directory, String fileName) throws RemoteException;
 	public KPFSFileInfo getFileLocation(String relPath) throws RemoteException;
-	public boolean addFileLocation(String relPath, String addr, int size) throws RemoteException;
+	public boolean addFileLocation(String relPath, String addr, long size) throws RemoteException;
 	public void removeFileLocation(String relPath, String addr) throws RemoteException;
 }
